top-down renderer to base
